403 error reported by remote access with auth
When auth is enabled, if the HTTPS request is not used when accessing the UI remotely, the client will not pass the access-token parameter for the websocket request, which will cause the interface to report a 403 error. Therefore, in order to avoid such problems, a new https control parameter is added to expose to the foreground business, so that auth can be used normally in non-https scenarios.